requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
pytest-cov>=4.1.0

# pip install -r requirements.txt